fix bad error message on eval with string argument
Using eval with a string argument previously failed because the code assumed
the first argument to eval would be a command--but of course it could be a
quoted sequence of commands and arguments.

I've backed this off to just let these through for now, but started throwing a
noisy warning nudging people to provide feedback on what they think the
normal behavior should be here.