pull patches from resholved src
The oil patches previously came from the directory the .nix files are
in. This restructures a little to pull them out of resholved's src and
pass them explicitly in to deps.nix. This should enable nixpkgs to
leverage the patches without having to use fetchPatch or having to
put all of the patches in nixpkgs and deal with keeping them in sync.