improve alias support, refactor resolution code
Supporting aliases well required a significant refactor of the resolution
code to give it a more durable sense of what context each command was
encountered in, what the relevant resolution order at that location was,
and how to resolve it according to the relevant order.

I also took the opportunity to do some needed cleanup.