abstract OSH AST use -> Invocation/Word wrappers
This is yet another sprawling refactor, but we get a lot for it.
resholve has been visiting the Oil/OSH AST and directly extracting
relevant information as it goes. I've been wanting to abstract this
for a while now, but had been hoping to put it off a little longer.

The underlying problem in #32 went ahead and forced the issue here,
so I've ripped off the bandaid and shifted all of the existing code
to the new abstraction.