add a special-case regex replacement for known aliases
Known aliases will appear somewhere in an `alias known_alias=def` form. When
`known_alias` is also a command in its own right, the current replacement
heuristics would also substitute the alias name (but note, this was a regression
caused during refactoring this past Summer, and missed because the tests were a
little too loose).

When we know a cmdlike is also an alias, we'll use a regex replacement routine
so that we can assert a smidge more control over which instances get substituted.

Fixes #72.