get script from unixtools instead of util-linux
See https://github.com/NixOS/nixpkgs/pull/232713 for full context, but
util-linux was an alias to unixtools on macOS, and true util-linux on
linux. resholve's tests unnecessarily depend on that detail to access
each platform's `script` command, and this broke when the linked PR
fixed compilation of util-linux on macOS.

`unixtools.script` will already point to the platform-specific tool.