Declare wordswurst’s rev and hash only once (#96)
The wordswurst package gets used in both make.nix and shell.nix. Before
this change, both of those files would call fetchFromGitHub separately.
In order to change the rev used for wordswurst, you would have to change
both files.

e7d851c (bump wordswurst, 2023-09-01) was supposed to update wordswurst,
but unfortunately it only updated the version of wordswurst that’s used
by make.nix. This change makes it so that both make.nix and shell.nix
use the same version of wordswurst so that we don’t run into this
problem in the future.