split command parser into located/unlocated
Existing CommandParser was intended for sussing out execed
args. These can, in some cases, ignore options that don't
affect its ability to do so. This rests on the assumption
that we've vetted all options for a command and know for
sure where all of the exec behavior is.

The changes are minimal, but I'm splitting it into a
LocatedExecCommandParser and UnlocatedExecCommandParser.
The former is equivalent to the old CommandParser. The
latter is meant to require strict syntax matching. It's a
bit of a straightjacket, but it enables people to carve out
syntax support for, say, just git add/rm/commit/reset but
save supporting the whole slate of options for later.