~fix conflict with abspaths that match a builtin
Some abspaths, such as /usr/bin/printf, could fall through the cracks
between the `fix` directive for re-resolving them and the builtin
check. Since builtins are higher priority, they'd fall into this code
path. But since builtins are basically no-ops from a replacement
perspective, the original abspath could end up left in.

This fix will treat these as if they should be externals, which is a
debatable choice. Feel free to object to the current handling in:

https://github.com/abathur/resholve/issues/111