improve cross situation
It might make sense to use nixpkgs normal mkDerivation infra
for some of these to minimize cross issues, but for now I
want to see if its acceptable to just use pkgsBuildBuild.

This also replaces our use of external sed with a sed parser
(note that awk is still using external awk; unsure whether
it has the same problem with hanging during cross as we ran
into with sed).