fix test break in bats 1.11
A shift in when/where bats executes the setup hook (it was previously
run at the start of each test; now it's run right before the start of
each test) causes a test that dynamically generates a script (in order
to embed an absolute path in the build sandbox) to be unable to find
the file at test time.

The key part of the fix is to generate the file in the test body. The
rest of this is just shortening paths and shifting to lean on per-test
tmpdirs bats already makes for us. These include the test number in
the path, making it marginally easier to debug test-specific issues if
they touch the tmpdir.

Note that tests/{behavior.bats,helpers.bash} carry some small ~quirks
for now in order to maintain compat with bats 1.10 and 1.11. This
probably isn't essential, but I'd rather have the bit of flexibility
and not need it than vice versa.