track doc edits in nixpkgs fcc95ff8172cc
This treewide nixpkgs commit is trying to force all nix blocks in md
files to be ~valid. Most of our examples are intentionally partial.
I think it's a bit misguided to change them in this case, but I agree
that making sure examples are valid is good and don't want to obstruct
that effort. The Nixpkgs copy of this is generated here and copied, so
I'll follow what they've done.