flake: support quick-and-dirty resolution testing
This feels a bit dirty, but we can build a script that in turn invokes
Nix a few times in order to support resolving with the Nix builder and
an invocation-time list of packages. (Hopefully flakes will have more
idiomatic support for passing these in at some point, but this works
for now I think?)

Intended to run like:
nix run .#resholve-with-packages -- <script-file> [<package>...]

Will output resolved script on stdout.