nix: configure and locally apply auto-formatting
Since everything in ./nixpkgs/ is shared with the nixpkgs repo, we
need to follow nixpkgs' decisions around formatting to minimize
friction. A previous commit tracked how this changed in nixpkgs, so
now this one sets up a formatter, applies formatting, and sets up
checks to ensure we stay sorted. Hopefully this won't churn much more.