test: skip blocked msmtp(q) parse tests
These tests started failing to actually invoke the underlying tool
after a bump of the version in nixpkgs. A ltitle more context in:
https://github.com/abathur/resholve/pull/103#issuecomment-3247459810

For now we'll skip them. Another option is not actually invoking these
tools in the process of testing them, but I'm generally of the mind
that requiring valid invocations will help ward off junk parse tests.