track addition of postResholve in nixpkgs pr 439997
This (currently-unmerged) PR adds a `postResholve` attr and
appends it to postFixup.

PR authored by dsg22 in 9bbf270466e43f2d23af34d2fff2c78c1e65c267

I will likely close that PR as soon as a I cut a resholve release
that includes this to let the bump PR supersede it.